+++ tests/unit/spec/src/messages/models/Container/insertMarkup.test.js
@@ -1,0 +1,42 @@
+import createContainer from 'utils/createContainer';
+import insertMarkup from 'src/messages/models/Container/insertMarkup';
+
+// JSDOM will not fire load events, causing insertMarkup to stall out
+HTMLImageElement.prototype.addEventListener = jest.fn((type, cb) => cb());
+
+describe('insertMarkup', () => {
+    afterEach(() => {
+        document.body.innerHTML = '';
+    });
+
+    it('Inserts markup string into iframe container', async () => {
+        const { container, getByText } = createContainer('iframe');
+        const markup = '<div><script>window.meta = { test: true };</script><div>test</div></div>';
+
+        await insertMarkup(container, markup);
+
+        expect(getByText(/test/i)).toBeVisible();
+    });
+
+    it('Inserts template node into iframe container', async () => {
+        const { container, getByText } = createContainer('iframe');
+        const markup = document.createElement('div');
+        markup.innerHTML = '<div><h1>test header</h1><p>see terms</p></div>';
+
+        await insertMarkup(container, markup);
+
+        expect(getByText(/test header/i)).toBeVisible();
+        expect(getByText(/see terms/i)).toBeVisible();
+    });
+
+    it('Inserts template node into div container', async () => {
+        const { container, getByText } = createContainer('div');
+        const markup = document.createElement('div');
+        markup.innerHTML = '<div><h1>test header</h1><p>see terms</p></div>';
+
+        await insertMarkup(container, markup);
+
+        expect(getByText(/test header/i)).toBeVisible();
+        expect(getByText(/see terms/i)).toBeVisible();
+    });
+});